Don't (un)link logical entities. Engine no likey..

+++ codemp/game/g_utils.cpp
@@ -1016,7 +1016,10 @@ void G_FreeEntity( gentity_t *ed ) {
 		return;
 	}
 
-	trap->UnlinkEntity ((sharedEntity_t *)ed);		// unlink from world
+	if ( !ed->isLogical )
+	{
+		trap->UnlinkEntity ((sharedEntity_t *)ed);		// unlink from world
+	}
 
 	trap->ICARUS_FreeEnt( (sharedEntity_t *)ed );	//ICARUS information must be added after this point
